Update Code.lisp
After losing my mind for a couple hours I realized I needed for the modifier to use copies instead of references which improved it greatly. Unfortunately there is still an underlying issue as the code reaches 85-95 fitness out of 100 very quickly and then seems to just start going downhill for 30 generations. Not sure why I'll look at it tomorrow.