Well it runs.....
Absolutely not good way of getting it to run, I just brute forced my way around double wrapping then unwrapping everything. Also includes a "root node" so the real root of the individuals can be swapped. Something about this code really really sucks when it comes to making a successful GP for the tree. Not sure if it's a logic bug or everything is wrong who knows