EVERYTHING WORKS!!!!!
absolutely scam coded ant evaluator (didn't use any of the functions progn2, progn3, if-food-ahead), and just put them in the evaluator function. I could not figure out how to balance function names, what is or isn't a macro, or eval. But who cares it's done

We just need the report stuff and maybe the printer for the map